==> analyses/molecular-subtyping-HGG/10-HGG-TP53-annotation.Rmd <==
---
title: "TP53 annotation for HGG"
author: "K S Gaonkar, Jo Lynne Rokita"
output: html_notebook
---

In this notebook we will annotate HGG samples with TP53 status we obtained from snv/cnv and TP53 classifier. We believe TP53 annotation will add useful information to the current known subtypes as we see TP53 mutations co-occurring with H3 mutations.

## Set up
```{r}
# Load in tidyverse functions
library(tidyverse)

root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
data_dir <- file.path(root_dir, "data")
# results folder
results_dir <- file.path(
  root_dir,
  "analyses",
  "molecular-subtyping-HGG",
  "results")
```


### Required files
```{r}
# File path to data directory
tp53_status_df <- read_tsv(file.path(
  root_dir,
  "analyses",
  "tp53_nf1_score",
  "results",
  "tp53_altered_status.tsv"), guess_max = 100000) %>%
  dplyr::rename(tp53_status = tp53_altered)

bs_match <- read_tsv(file.path(data_dir, "histologies-base.tsv"), guess_max = 100000) %>%
  filter(!is.na(pathology_diagnosis),
         !cohort %in% c("TCGA", "GTEx"))

hgg_subtypes <- read_tsv(file.path(results_dir,"HGG_cleaned_all_table.tsv")) %>%
  left_join(bs_match)

```

# Add tp53 status 
```{r}
hgg_subtypes_tp53 <- hgg_subtypes %>%
  left_join(unique(tp53_status_df[,c("match_id", "tp53_status")])) %>%
  mutate(molecular_subtype= case_when(
    tp53_status %in% c("activated", "loss") ~ stringr::str_c(molecular_subtype,", TP53"),
    TRUE ~ molecular_subtype
  )) %>%
  select(match_id, molecular_subtype, molecular_subtype_methyl, CNS_region, tp53_status) %>%
  unique() %>%
  group_by(match_id, molecular_subtype) %>%
  summarize(across(everything(), ~paste(unique(na.omit(.x)), collapse = ", ")), .groups = 'drop') %>% 
  right_join(hgg_subtypes[,c("Kids_First_Participant_ID", "Kids_First_Biospecimen_ID", "match_id", "sample_id")]) %>%
  arrange(match_id) %>%
  unique() %>%
  write_tsv(file.path(results_dir,"HGG_molecular_subtype_tp53.tsv"))

# check for duplicates
nrow(hgg_subtypes_tp53)-length(unique(hgg_subtypes_tp53$Kids_First_Biospecimen_ID))

hgg_subtypes_tp53[duplicated(hgg_subtypes_tp53$Kids_First_Biospecimen_ID) | duplicated(hgg_subtypes_tp53$Kids_First_Biospecimen_ID, fromLast = TRUE), ]

```
